Refactor CLI prompt orchestration
